server api - fix duplicate code issue
