+++ server/api/cpuLogs.js
@@ -1,21 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const Cpu = require('../models/cpu.js')
-
-// String attributes list to check with req.query object
-const stringAttributes = ['orderBy', 'deviceId', 'timestamp']
-
-// Number attributes list to check with req.query object
-const numberAttributes = [
-  'limit',
-  'orderValue',
-  'usagePercentage',
-  'usageSpeed',
-  'numProcesses',
-  'threadsAlive',
-  'threadsSleeping',
-  'uptime',
-]
+const filterData = require('./shared/filter')
 
 // get X number of entries for single device (limit, deviceId)
 router.get('/specific-device', async (req, res) => {
@@ -74,35 +60,7 @@ router.get('/timestamp', async (req, res) => {
 // get multiple entries given an attribute with the ability to add a limit and order by filter
 router.get('/specific-attribute', async (req, res) => {
   try {
-    let options = {}
-    let query = {}
-    let queryObj = Object(req.query)
-    for (let k in queryObj) {
-      queryObj[k] = queryObj[k].split(',')
-      if (stringAttributes.includes(k)) {
-        query[String(k)] = queryObj[k]
-      }
-      if (numberAttributes.includes(k)) {
-        query[String(k)] = queryObj[k].map(Number)
-      }
-    }
-    if (query.limit) {
-      options.limit = query.limit
-      delete query.limit
-    }
-    if (query.orderBy) {
-      var orderBy = query.orderBy
-      delete query.orderBy
-    }
-    if (query.orderValue) {
-      var orderValue = query.orderValue
-      delete query.orderValue
-    }
-    if (orderValue && orderBy) {
-      options.sort = {
-        [orderBy]: orderValue,
-      }
-    }
+    let query = filterData(req.query)
     await Cpu.CpuLogs.find({ $and: [query] }, {}, options).exec(
       (err, cpuLogs) => {
         return res.status(200).json(cpuLogs)
