+++ server/api/wifiLogs.js
@@ -1,18 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const Wifi = require('../models/wifi.js')
-
-// String attributes list to check with req.query object
-const stringAttributes = ['orderBy', 'deviceId', 'timestamp']
-
-// Number attributes list to check with req.query object
-const numberAttributes = [
-  'limit',
-  'orderValue',
-  'sendSpeed',
-  'receiveSpeed',
-  'signalStrength',
-]
+const filterData = require('./shared/filter')
 
 // get X number of entries for single device (limit, deviceId)
 router.get('/specific-device', async (req, res) => {
@@ -71,35 +60,7 @@ router.get('/timestamp', async (req, res) => {
 // get multiple entries given an attribute with the ability to add a limit and order by filter
 router.get('/specific-attribute', async (req, res) => {
   try {
-    let options = {}
-    let query = {}
-    let queryObj = Object(req.query)
-    for (let k in queryObj) {
-      queryObj[k] = queryObj[k].split(',')
-      if (stringAttributes.includes(k)) {
-        query[String(k)] = queryObj[k]
-      }
-      if (numberAttributes.includes(k)) {
-        query[String(k)] = queryObj[k].map(Number)
-      }
-    }
-    if (query.limit) {
-      options.limit = query.limit
-      delete query.limit
-    }
-    if (query.orderBy) {
-      var orderBy = query.orderBy
-      delete query.orderBy
-    }
-    if (query.orderValue) {
-      var orderValue = query.orderValue
-      delete query.orderValue
-    }
-    if (orderValue && orderBy) {
-      options.sort = {
-        [orderBy]: orderValue,
-      }
-    }
+    let query = filterData(req.query)
     await Wifi.WifiLogs.find({ $and: [query] }, {}, options).exec(
       (err, wifiLogs) => {
         return res.status(200).json(wifiLogs)
